requirements.txt
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
zipp==3.8.0
